% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dur_sim.R
\docType{data}
\name{dur_sim}
\alias{dur_sim}
\title{dur_sim}
\format{
\subsection{\code{dur_sim}}{
}
}
\usage{
dur_sim
}
\description{
A simulated dataset for illustrating clusdur functionalities
}
\keyword{datasets}
