% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{long_to_wide}
\alias{long_to_wide}
\title{long_to_wide}
\usage{
long_to_wide(data, id)
}
\arguments{
\item{data}{The timestamp data in long format}

\item{id}{Respondent identifier}
}
\value{
A data.table object holding the relative ranks of durations in wide format
}
\description{
Reshapes duration data from long to wide for cluster analysis
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
