% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusdur.R
\name{plot.clusdur}
\alias{plot.clusdur}
\title{plot.clusdur}
\usage{
\method{plot}{clusdur}(fit)
}
\arguments{
\item{fit}{A clusdur object}
}
\value{
A ggplot object`
}
\description{
Provides an overview table for the time and scope conditions of
a data set
}
\examples{
# load simulated data
data(dur_sim)

# fit cluster analysis
fit <- clusdur(data = dur_sim, id = "id", cols = 2:31, method = "ward.D2", k = 2)

# plot distribution across deciles
plot(fit)
}
