% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_clusdur.R
\name{shiny_clusdur}
\alias{shiny_clusdur}
\title{shiny_clusdur}
\usage{
shiny_clusdur(fit, facet = TRUE)
}
\arguments{
\item{fit}{A "clusdur" object}
}
\value{
A shiny app showing the distribution of durations faceted by cluster membership.
}
\description{
A basic shiny app that shows the duration distribution for all columns by cluster.
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
