% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wide_to_long}
\alias{wide_to_long}
\title{wide_to_long}
\usage{
wide_to_long(data, id, cols)
}
\arguments{
\item{data}{The timestamp data in wide format}

\item{id}{Respondent identifier}

\item{cols}{The columns holding the duration data}
}
\value{
A data.table object containing the columns id, item, duration, duration_rel, duration_decile
}
\description{
Reshapes duration data from wide to long and generate relative ranks
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
