% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusdur.R
\name{clusdur}
\alias{clusdur}
\title{clusdur}
\usage{
clusdur(data, id, cols, method = "ward.D2", k = 2)
}
\arguments{
\item{data}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{cols}{Time (e.g., time periods are given by years, months, ...)}

\item{method}{The linkage method (one of "complete", "ward.D2", "average)}

\item{k}{Number of clusters}
}
\value{
A clusdur object
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
