% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusdur.R
\name{plot.clusdur}
\alias{plot.clusdur}
\title{plot.clusdur}
\usage{
\method{plot}{clusdur}(fit)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
a data set
}
\examples{
# load simulated data
data(dur_sim)

# fit cluster analysis
fit <- clusdur(data = dur_sim, id = "id", cols = 2:51, method = "ward.D2", k = 2)

# plot distribution across deciles
plot(fit)
}
