% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{wide_to_long}
\alias{wide_to_long}
\title{wide_to_long}
\usage{
wide_to_long(data, id, cols)
}
\arguments{
\item{id}{Scope (e.g., country codes or individual IDs)}

\item{dat}{A data set object}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
